Simplify naming of settings component.
