+++ extension/src/settings/Settings.scss
